+++ packages/bigquery/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-bigquery
 
+## 4.0.7 - 02 December 2025
+
+### Patch Changes
+
+- e8f7f2d: Security update
+- 6aa9800: Security update
+
 ## 4.0.6 - 28 November 2025
 
 ### Patch Changes