+++ packages/primero/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-primero
 
+## 4.0.7 - 02 December 2025
+
+### Patch Changes
+
+- 6aa9800: Security update
+
 ## 4.0.6 - 28 November 2025
 
 ### Patch Changes